feat(SPV-1394) Update Admin API options
feat(admin-api): add AdminClientOptions for admin API
docs(readme): update docs with admin and client separation description
test(examples): update example test files with new admin api changes
chore(package.json): update beta versioning to 3.2
