+++ examples/admin-remove-user.ts
@@ -1,5 +1,5 @@
 import { SPVWalletAdminAPI } from '../dist/typescript-npm-package.cjs.js';
-import { exampleAdminKey, examplePaymail } from './example-keys.js';
+import { exampleAdminKey } from './example-keys.js';
 import { errMessage } from './utils.js';
 
 const server = 'http://localhost:3003';
@@ -9,6 +9,6 @@ if (!exampleAdminKey) {
   process.exit(1);
 }
 
-const adminClient = new SPVWalletAdminAPI(server, exampleAdminKey);
+const adminClient = new SPVWalletAdminAPI(server, {adminKey : exampleAdminKey});
 
-await adminClient.deletePaymail(examplePaymail, examplePaymail);
+await adminClient.deletePaymail("d43ed481ba08aae1db02d880ebefe962f9796168387bb293a95024cb02b953ef");
