+++ examples/package.json
@@ -11,16 +11,16 @@
     "admin-add-user": "yarn run:example admin-add-user.ts",
     "admin-remove-user": "yarn run:example admin-remove-user.ts",
     "admin-get-handled-paymail-domains": "yarn run:example admin-get-handled-paymail-domains.ts",
-    "user-generate-totp": "yarn run:example generate-totp.ts",
-    "user-access-key": "yarn run:example access-key.ts",
-    "user-get-balance": "yarn run:example get-balance.ts",
+    "manage-contacts": "yarn run:example manage-contacts.ts",
+    "access-key": "yarn run:example access-key.ts",
+    "get-balance": "yarn run:example get-balance.ts",
     "user-get-handled-paymail-domains": "yarn run:example user-get-handled-paymail-domains.ts",
-    "user-sync-merkleroots": "yarn run:example sync-merkleroots.ts",
+    "sync-merkleroots": "yarn run:example sync-merkleroots.ts",
     "user-update-xpub-metadata": "yarn run:example update-user-xpub-metadata.ts",
-    "user-get-shared-config": "yarn run:example get-shared-config.ts",
-    "user-list-transactions": "yarn run:example list-transactions.ts",
-    "user-create-transaction": "yarn run:example create-transaction.ts",
-    "user-send-op-return": "yarn run:example send-op-return.ts",
+    "get-shared-config": "yarn run:example get-shared-config.ts",
+    "list-transactions": "yarn run:example list-transactions.ts",
+    "create-transaction": "yarn run:example create-transaction.ts",
+    "send-op-return": "yarn run:example send-op-return.ts",
     "xpriv-from-mnemonic": "yarn run:example xpriv-from-mnemonic.ts",
     "xpub-from-xpriv": "yarn run:example xpub-from-xpriv.ts",
     "generate-keys": "yarn run:example generate-keys.ts",
