+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bsv/spv-wallet-js-client",
-  "version": "1.0.0-beta.31",
+  "version": "1.0.0-beta.32",
   "description": "TypeScript library for connecting to a SPV Wallet server",
   "repository": {
     "type": "git",
